+++ structs/db/DB_LogRiskControl.go
@@ -6,7 +6,7 @@ type DB_LogRiskControl struct {
 	LId  int    `json:"LId" gorm:"column:触发在线id;"`
 	User string `json:"User" gorm:"column:User;size:200;index;comment:用户名"`
 	Type int    `json:"Type" gorm:"column:Type;comment:风控规则类型"`
-	Time int    `json:"Time" gorm:"column:Time;comment:时间"`
+	Time int64  `json:"Time" gorm:"column:Time;comment:时间"`
 	Ip   string `json:"Ip" gorm:"column:Ip;size:200;comment:ip地址"`
 	Note string `json:"Note" gorm:"column:Note;size:10000;comment:风控信息"`
 }
